Fixed small type error
I was doing computations with doubles but stored the result as float.
It has been fixed.

+++ RingBuffer/RingBuffer/RingBuffer.cpp
@@ -37,7 +37,7 @@ double RB::AverageContents()
 	{
 		sum += buffer[i];
 	}
-	float avg = static_cast<double>(sum) / static_cast<double>(capacity);
+	double avg = static_cast<double>(sum) / static_cast<double>(capacity);
 	return avg;
 }
